+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 English | [简体中文](./CHANGELOG_CN.md)
 
+## 3.10.1 (2021-12-23)
+
+- `Feat(Network)` Pretty output format for JSON response. (issue #486)
+- `Fix(Style)` Avoid panel scaling with the web page When `initial-scale !== 1`.
+- `Fix(Core)` Fix the issue that clicking the panel button does not work in PC mode. (issue #487)
+- `Fix(Network)` Display formatted JSON instead of `[object Object]` when Query/Payload/Headers is an object or array.
+- `Fix(Network)` Avoid overwriting `onreadystatechange` of XHR objects multiple times when XHR objects are reused in some cases. (issue #214)
+
+
 ## 3.10.0 (2021-12-17)
 
 > In this version, we refactored a lot of core logic, and used Svelte as the rendering engine for all views.